refactor: Move `src/store.ts` to `src/redux/store.ts`
