+++ src/redux/ConnectorState/CurrentConnectorSlice.ts
@@ -1,6 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 
-import type { RootState } from '/store'
+import type { RootState } from '/redux/store'
 
 export interface CurrentConnectorState {
   currentRunningConnector?: string
